# expression: implement vectorized evaluation for `builtinWeekWithoutModeSig`