--- a/expression/builtin_time_vec.go
+++ b/expression/builtin_time_vec.go
@@ -14,13 +14,13 @@
 package expression
 
 import (
-	"github.com/pingcap/tidb/sessionctx/variable"
 	"math"
 	"strconv"
 	"time"
 
 	"github.com/pingcap/errors"
 	"github.com/pingcap/parser/mysql"
+	"github.com/pingcap/tidb/sessionctx/variable"
 	"github.com/pingcap/tidb/types"
 	"github.com/pingcap/tidb/util/chunk"
 )